Moved spectral match results visualizer to own package in vis/spectra
