+++ src/main/java/net/sf/mzmine/modules/visualization/spectra/spectralmatchresults/SpectralIdentificationSpectralDatabaseTextPane.java
@@ -15,7 +15,7 @@
  * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
  * USA
  */
-package net.sf.mzmine.modules.visualization.spectra.simplespectra.spectraidentification.spectraldatabase;
+package net.sf.mzmine.modules.visualization.spectra.spectralmatchresults;
 
 import javax.swing.JTextPane;
